requirements.txt
psycopg2-binary
pandas
numpy
matplotlib
reportlab
pdfkit
argparse
rich
bcrypt
